Rename `Location` to `Point` and `Region` to `Location`
