+++ src/parser/location.rs
@@ -1,138 +1,84 @@
-use lazy_static::lazy_static;
-use regex::Regex;
-use std::fmt::{self, Display};
+use core::fmt::{self, Display};
+use crate::parser::point::Point;
 
-lazy_static! {
-    static ref NEWLINE: Regex = Regex::new("(\n|\r\n|\r)").unwrap();
+#[derive(Clone, Debug)]
+pub struct Location<'i> {
+    file_name: &'i str,
+    src: &'i str,
+    lines: (usize, usize),
+    cols: (usize, usize),
+    indices: (usize, usize),
 }
 
-#[derive(Clone, Debug, Default)]
-pub struct Location<'input> {
-    pub file_name: &'input str,
-    pub src: &'input str,
-    pub line: usize,
-    pub col: usize,
-    pub index: usize,
-}
-
-impl<'input> Location<'input> {
-    pub fn new(fname: &'input str, src: &'input str) -> Self {
-        Self {
-            file_name: fname,
-            src,
-            index: 0,
-            line: 1,
-            col: 0,
-        }
-    }
-
-    pub fn shift(mut self, text: &'input str) -> Self {
-        let lines: Vec<&str> = NEWLINE.split(text).into_iter().collect();
-        let num_lines = lines.len();
-
-        self.line += num_lines - 1;
-
-        let last_line = lines[num_lines - 1];
-        let last_line_width = last_line
-            .chars()
-            .map(|c| if c == '\t' { 4 } else { 1 })
-            .sum();
-        self.col = if num_lines > 1 {
-            last_line_width
-        } else {
-            self.col + last_line_width
-        };
-
-        self.index += text.len();
-
-        self
-    }
-
+impl<'i> Location<'i> {
     #[allow(dead_code)]
-    pub fn text_upto(&self, other: &Location) -> Option<&'input str> {
-        if self.file_name != other.file_name {
-            return None;
+    pub fn new(start: &Point<'i>, end: &Point<'i>) -> Self {
+        Self {
+            file_name: start.file_name,
+            src: start.src,
+            lines: (start.line, end.line),
+            indices: (start.index, end.index),
+            cols: (start.col, end.col),
         }
-        Some(&self.src[self.index..other.index])
+    }
+
+    pub fn file_name(&self) -> &str {
+        self.file_name
+    }
+
+    pub fn src(&self) -> &str {
+        self.src
+    }
+
+    pub fn lines(&self) -> (usize, usize) {
+        self.lines
+    }
+
+    pub fn cols(&self) -> (usize, usize) {
+        self.cols
+    }
+
+    pub fn indices(&self) -> (usize, usize) {
+        self.indices
     }
 }
 
-impl<'input> Display for Location<'input> {
+impl Display for Location<'_> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}:{}:({})", self.file_name, self.line, self.index)
+        if self.lines.0 != self.lines.1 {
+            write!(
+                f,
+                "{}:{}-{}:{}-{}",
+                self.file_name, self.lines.0, self.lines.1, self.cols.0, self.cols.1
+            )
+        } else if self.cols.0 != self.cols.1 {
+            write!(
+                f,
+                "{}:{}:{}-{}",
+                self.file_name, self.lines.0, self.cols.0, self.cols.1
+            )
+        } else {
+            write!(f, "{}:{}:{}", self.file_name, self.lines.0, self.cols.0)
+        }
     }
 }
 
 #[cfg(test)]
 mod test {
     use super::*;
+
     #[test]
     fn new() {
-        let src = "content";
-        let loc = Location::new("fname", src);
+        let text = "my name\nis methos";
+        let start = Point::new("fname.em", text);
+        let end = start.clone().shift(text);
 
-        assert_eq!("fname", loc.file_name);
-        assert_eq!(src, loc.src);
-        assert_eq!(0, loc.index);
-        assert_eq!(1, loc.line);
-        assert_eq!(0, loc.col);
-    }
+        let region = Location::new(&start, &end);
 
-    #[test]
-    fn shift_single_line() {
-        let src = "my name is methos";
-        let start = Location::new("fname", src);
-        let mid = start.clone().shift("my name is ");
-        let end = mid.clone().shift("methos");
-
-        assert_eq!("fname", mid.file_name);
-        assert_eq!(src, mid.src);
-        assert_eq!(11, mid.index);
-        assert_eq!(1, mid.line);
-        assert_eq!(11, mid.col);
-
-        assert_eq!("fname", end.file_name);
-        assert_eq!(src, end.src);
-        assert_eq!(17, end.index);
-        assert_eq!(1, end.line);
-        assert_eq!(17, end.col);
-
-        assert_eq!(Some("my name is "), start.text_upto(&mid));
-        assert_eq!(Some("methos"), mid.text_upto(&end));
-        assert_eq!(Some("my name is methos"), start.text_upto(&end));
-    }
-
-    #[test]
-    fn tabs() {
-        let src = "\thello,\tworld";
-        let start = Location::new("fname", src);
-        let end = start.shift(src);
-
-        assert_eq!(13, end.index);
-        assert_eq!(19, end.col);
-    }
-
-    #[test]
-    fn shift_multi_line() {
-        let raw_src = "Welcome! Welcome to City 17! You have chosen, or been chosen, to relocate to one of our finest remaining urban centres";
-        let src = raw_src.replace(" ", "\n");
-        let start = Location::new("file_name", &src);
-        let end = start.clone().shift(&src);
-
-        assert_eq!("file_name", end.file_name);
-        assert_eq!(src, end.src);
-        assert_eq!(21, end.line);
-        assert_eq!(118, end.index);
-        assert_eq!(7, end.col);
-
-        assert_eq!(start.text_upto(&end), Some(&src[..]));
-    }
-
-    #[test]
-    fn text_upto_different_locations() {
-        let l1 = Location::new("file1", "fubar");
-        let l2 = Location::new("file2", "snafu");
-
-        assert_eq!(l1.text_upto(&l2), None);
+        assert_eq!("fname.em", region.file_name());
+        assert_eq!(text, region.src());
+        assert_eq!((start.line, end.line), region.lines());
+        assert_eq!((start.col, end.col), region.cols());
+        assert_eq!((start.index, end.index), region.indices());
     }
 }
